+++ client/app/components/Chart/index.jsx
@@ -1,11 +1,8 @@
 // @flow
-/* eslint react/jsx-props-no-spreading: 0 */
-import { Chart as ChartJS } from 'chart.js';
-import React from 'react';
+import React, { Suspense, lazy } from 'react';
 import type { Node } from 'react';
-import { AreaChart, LineChart } from 'react-chartkick';
 
-ChartJS.defaults.global.defaultFontFamily = 'Lato';
+const ChartRenderer = lazy(() => import('./ChartRenderer'));
 
 type chartShape = {
   xtitle?: string,
@@ -14,12 +11,10 @@ type chartShape = {
   chartType: 'Line' | 'Area',
 };
 
-const colorSchemes = ['#6D0839', '#66118', '#7F503F', '#775577', '#CCAADD'];
-
-export function Chart({ chartType, ...props }: chartShape): Node {
-  return chartType === 'Line' ? (
-    <LineChart {...props} colors={colorSchemes} />
-  ) : (
-    <AreaChart {...props} colors={colorSchemes} />
+export function Chart(props: chartShape): Node {
+  return (
+    <Suspense fallback={null}>
+      <ChartRenderer {...props} />
+    </Suspense>
   );
 }
